# web/api-spec.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "ServerC",
    "version": "1.0"
  },
  "paths": {
...
    "/Features": {
      "post": {
        "tags": ["Features"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFeatureInput"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFeatureInput"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFeatureInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Feature"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Feature"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Feature"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Features"],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Feature"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Feature"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Feature"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Features"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Feature"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Feature"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Feature"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Feature"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Feature"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Feature"
                }
              }
            }
          }
        }
      }
    },
    "/Features/company/{companyId}": {
      "get": {
        "tags": ["Features"],
        "parameters": [
          {
            "name": "companyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Feature"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Feature"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Feature"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/Features/{featureId}": {
      "delete": {
        "tags": ["Features"],
        "parameters": [
          {
            "name": "featureId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateFeatureInput": {
        "required": ["companyID", "featureName"],
        "type": "object",
        "properties": {
          "featureName": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "companyID": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "Feature": {
        "required": ["companyID", "featureName"],
        "type": "object",
        "properties": {
          "featureID": {
            "type": "integer",
            "format": "int32"
          },
          "featureName": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "companyID": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}