# web/generated-api/models/CreateFeatureInput.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ServerC
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CreateFeatureInput
 */
export interface CreateFeatureInput {
    /**
     * 
     * @type {string}
     * @memberof CreateFeatureInput
     */
    featureName: string;
    /**
     * 
     * @type {number}
     * @memberof CreateFeatureInput
     */
    companyID: number;
}

/**
 * Check if a given object implements the CreateFeatureInput interface.
 */
export function instanceOfCreateFeatureInput(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "featureName" in value;
    isInstance = isInstance && "companyID" in value;

    return isInstance;
}

export function CreateFeatureInputFromJSON(json: any): CreateFeatureInput {
    return CreateFeatureInputFromJSONTyped(json, false);
}

export function CreateFeatureInputFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreateFeatureInput {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'featureName': json['featureName'],
        'companyID': json['companyID'],
    };
}

export function CreateFeatureInputToJSON(value?: CreateFeatureInput | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'featureName': value.featureName,
        'companyID': value.companyID,
    };
}